lvm: messages updates
Add missing periods in _split_mirror_images error messages.
Avoid passing NULL layer to log_error in dm_build_dm_name.
Avoid passing NULL to %s format specifier.
Cast to match format specifier.
Demote first-interval message from log_error to log_debug.
Fix "he only" typo in comment.
Fix "Internal user only" typo in VDO flag comments.
Fix "to high" typo in error message.
Fix %d format specifier for unsigned expression.
Fix always-constant log message in lockd_lvcreate_done.
Fix copy-paste "Suspend" to "Resume" in _resume_node deptree.
Fix copy-paste error in close() error message.
Fix copy-paste error in lock_args hash error message.
Fix copy-paste error in log_error and formatting.
Fix copy-paste error in lvmcache_is_outdated_dev log message.
Fix double space and %d for unsigned in error message.
Fix error message "debug" to "foreground" for argument parsing.
Fix error message in find_mirror_seg.
Fix error message to say precommitted instead of preload metadata.
Fix format specifier for dm_list_size unsigned return.
Fix format specifier for uint32_t.
Fix format specifier for uint64_t in simple tree dump.
Fix format specifier for unsigned status type.
Fix incorrect comment about lvs_in_vg return values.
Fix inverted "found" to "not found" in error message.
Fix misleading coverity comment for memset sentinel.
Fix misleading VG error message.
Fix missing NULL guard for params in dev_manager_thin_device_id.
Fix missing NULL guard in dev_manager_thin_device_id error message.
Fix missing space in stripe_filler log messages.
Fix missing verb in error message.
Fix truncated function name in _merge_free_pv_segment error message.
Fix typo "sem_ctl" to "semctl" in error message.
Fix typo "to short" to "too short" in error message.
Fix typo in lockd_start_vg debug message.
Fix typo in vg_remove_snapshot comment.
Fix wrong extent in lv_split_segment error messages.
Fix wrong function name and missing period in error message.
Fix wrong function name in error message.
Fix wrong function name in INTERNAL_ERROR message.
Fix wrong variable in segment reduction error message.
Fix wrong variable in _node_has_closed_parents log message.
Remove misleading WARNING prefix from log_error message.
Remove spurious "bytes" from cache_vol_set_params error message.
Use display_lvname in data_lv_from_thin_pool error message.

+++ lib/cache_segtype/cache.c
@@ -697,7 +697,7 @@ static int _cache_add_target_line(struct dev_manager *dm,
 						if (!_accepted[i].settings[j]) {
 							log_warn("WARNING: %s cache policy does not support \"%s=" FMTu64 "\" setting, "
 								 "remove with 'lvchange --cachesettings \"%s=default\" ...'.",
-								 _accepted[i].name, cn->key, cn->v->v.i, cn->key);
+								 _accepted[i].name, cn->key, (uint64_t)cn->v->v.i, cn->key);
 							dm_config_remove_node(policy_settings, cn);
 							goto restart;
 						}
